requirements.txt
discord.py==0.16.12
discord.py[voice] == 0.16.0
aiohttp
websockets
chardet
requests
praw
asyncio
json
